Committing the files to support Restful service for patient registration.

+++ src/main/java/com/wenzins/webhis/dao/PatientDAO.java
@@ -9,5 +9,5 @@ public interface PatientDAO {
 
 	public List<Patient> listAllPatients();
 
-	public void deletePatientDetails(Integer patientId);
+	public void deletePatientDetails(long patientId);
 }
